+++ Include/Acidanthera/Library/OcConfigurationLib.h
@@ -575,9 +575,9 @@ typedef enum {
 ///
 #define OC_UEFI_AUDIO_FIELDS(_, __) \
   _(OC_STRING                   , AudioDevice        ,     , OC_STRING_CONSTR ("", _, __)  , OC_DESTR (OC_STRING)) \
+  _(OC_STRING                   , PlayChime          ,     , OC_STRING_CONSTR ("", _, __)  , OC_DESTR (OC_STRING)) \
   _(UINT16                      , VolumeAmplifier    ,     , 0                             , ()) \
   _(BOOLEAN                     , AudioSupport       ,     , FALSE                         , ()) \
-  _(BOOLEAN                     , PlayChime          ,     , FALSE                         , ()) \
   _(UINT8                       , AudioCodec         ,     , 0                             , ()) \
   _(UINT8                       , AudioOut           ,     , 0                             , ()) \
   _(UINT8                       , MinimumVolume      ,     , 0                             , ())